refactor: break up cache logic into separate files
